+++ app/classes/ChargeData.php
@@ -70,8 +70,9 @@ class ChargeData {
 		foreach ($headers as $header) {
 			$ths[] = $Html->th($header);
 		}
+
 		$trs[] = $Html->tr(implode('', $ths));
-		$totales = array('valor_tarificada' => 0, 'monto_cobrado' => 0);
+		$totales = array('valor_tarificada' => 0, 'monto_cobrado' => 0, 'duracion' => 0, 'duracion_cobrada' => 0, 'duracion_incobrables' => 0);
 		$currency_columns = array_keys($totales);
 		foreach ($data as $row) {
 			$tds = array();
@@ -160,11 +161,6 @@ class ChargeData {
 		return $this->matter_sumary[$matter_code];
 	}
 
-	public function getSumaryAndTotalOrderedBy($sql_order_by) {
-		$works = $this->getWorksOrdererBy($sql_order_by);
-		return $this->getSumaryAndTotal($works);
-	}
-
 	public function getSumaryByCategory() {
 		$this->sumaryByCategory();
 		return $this->category_sumary;
@@ -205,15 +201,10 @@ class ChargeData {
 		return $this->get("opc_{$opt}") == 1;
 	}
 
-
 	/**
 	 * Cartga todos los trabajos del cobro
 	 */
 	protected function loadWorks() {
-		$this->works = $this->getWorksOrdererBy();
-	}
-
-	public function getWorksOrdererBy($sql_order_by = 'trabajo.fecha ASC,  trabajo.descripcion ASC') {
 		$time_format = '%k:%i';
 		$Criteria = new Criteria($this->Sesion);
 		$Criteria->add_from('trabajo')
@@ -223,8 +214,8 @@ class ChargeData {
 			->add_select('trabajo.descripcion')
 			->add_select('trabajo.fecha')
 			->add_select('trabajo.id_usuario')
-			->add_select('trabajo.visible')
 			->add_select('trabajo.cobrable')
+			->add_select('trabajo.visible')
 			->add_select('trabajo.id_trabajo')
 			->add_select('trabajo.tarifa_hh')
 			->add_select('IF(trabajo.cobrable, trabajo.tarifa_hh * (TIME_TO_SEC(trabajo.duracion_cobrada) / 3600 ), 0)', 'importe')
@@ -252,11 +243,12 @@ class ChargeData {
 		$Criteria = $this->scopeUserCategory($Criteria);
 		$Criteria = $this->scopeChargeable($Criteria);
 
-		$Criteria->add_ordering($sql_order_by);
-
-		$works = $Criteria->run();
-		foreach ($works as $i => $work) {
+		$this->works = $Criteria->run();
+		foreach ($this->works as $i => $work) {
 			$work['duracion'] = Utiles::GlosaHora2Multiplicador($work['glosa_duracion']);
+			$work['duracion_incobrables'] = $work['cobrable'] ? 0 : $work['duracion'];
+			$work['glosa_duracion_incobrables'] = $work['cobrable'] ? 0 : $work['glosa_duracion'];
+
 			$work['duracion_cobrada'] = Utiles::GlosaHora2Multiplicador($work['glosa_duracion_cobrada']);
 			$work['duracion_retainer'] = Utiles::GlosaHora2Multiplicador($work['glosa_duracion_retainer']);
 			$work['duracion_descontada'] = $work['duracion'] - $work['duracion_cobrada'] - $work['duracion_incobrables'];
@@ -269,8 +261,7 @@ class ChargeData {
 			} else  if ($this->get('forma_cobro') == 'ESCALONADA') {
 				$work['monto_cobrado_escalonada'] = $work['monto_cobrado'];
 			}
-			// WHY?: los incobrables
-			//$work['duracion_tarificada'] = $sumary[$user_id]['duracion_cobrada'] - $sumary[$user_id]['duracion_incobrables'];
+
 			$work['duracion_tarificada'] = max($work['duracion_cobrada'] - $work['duracion_retainer'], 0);
 			$work['valor_tarificada'] = $work['duracion_tarificada'] * $work['tarifa_hh'];
 
@@ -284,9 +275,8 @@ class ChargeData {
 				$work['glosa_duracion_retainer'] = Utiles::Decimal2GlosaHora($work['duracion_retainer']);
 			}
 
-			$works[$i] = $work;
+			$this->works[$i] = $work;
 		}
-		return $works;
 	}
 
 	/**
@@ -295,7 +285,25 @@ class ChargeData {
 	 * @return Criteria
 	 */
 	protected function scopeUserCategory(Criteria $Criteria) {
-		$Criteria->add_select('prm_categoria_usuario.id_categoria_usuario');
+		if (Conf::read('OrdenarPorCategoriaNombreUsuario') || Conf::read('OrdenarPorCategoriaUsuario')) {
+			$Criteria->add_select('prm_categoria_usuario.id_categoria_usuario')
+ 				->add_ordering('prm_categoria_usuario.orden')
+ 				->add_ordering('usuario.id_usuario');
+		} else if (Conf::read('SepararPorUsuario')) {
+			$Criteria->add_select('prm_categoria_usuario.id_categoria_usuario')
+				->add_ordering('usuario.id_categoria_usuario')
+				->add_ordering('usuario.id_usuario');
+		} else if (Conf::read('OrdenarPorCategoriaDetalleProfesional')) {
+			$Criteria->add_ordering('usuario.id_categoria_usuario', 'DESC');
+		} else if (Conf::read('OrdenarPorFechaCategoria')) {
+			$Criteria->add_select('prm_categoria_usuario.id_categoria_usuario')
+				->add_ordering('trabajo.fecha')
+				->add_ordering('usuario.id_categoria_usuario')
+				->add_ordering('usuario.id_usuario');
+		} else {
+			$Criteria->add_ordering('trabajo.fecha', 'ASC')
+				->add_ordering('trabajo.descripcion');
+ 		}
  		if ($this->get('codigo_idioma') == 'es') {
  			$Criteria->add_select('prm_categoria_usuario.glosa_categoria', 'categoria');
  		} else {
@@ -310,31 +318,32 @@ class ChargeData {
 	 * @return Criteria
 	 */
 	protected function scopeChargeable(Criteria $Criteria) {
-		if ($this->opt('ver_horas_trabajadas')) {
-			$Criteria->add_restriction(Restriction::equals('trabajo.cobrable', '1'));
-		}
-
-		if (!$this->opt('ver_cobrable')) {
-			return $Criteria->add_restriction(Restriction::equals('trabajo.visible', '1'));
-		}
-
-		return $Criteria->add_restriction(
-			Restriction::or_clause(
-				Restriction::and_clause(
-					Restriction::equals('trabajo.cobrable', '0'),
-					Restriction::equals('trabajo.visible', '0')
-				),
-				Restriction::and_clause(
-					Restriction::equals('trabajo.cobrable', '1'),
-					Restriction::equals('trabajo.visible', '1')
-				)
+		$conditions = [
+			Restriction::and_clause(
+				Restriction::equals('trabajo.cobrable', 1),
+				Restriction::equals('trabajo.visible', 1)
 			)
-		);
+		];
 
+		if ($this->opt('ver_horas_trabajadas')) {
+			$conditions[] = Restriction::and_clause(
+				Restriction::equals('trabajo.cobrable', 0),
+				Restriction::equals('trabajo.visible', 1)
+			);
+		}
+
+		if ($this->opt('ver_cobrable')) {
+			$conditions[] = Restriction::and_clause(
+				Restriction::equals('trabajo.cobrable', 0),
+				Restriction::equals('trabajo.visible', 0)
+			);
+		}
+		$Criteria->add_restriction(Restriction::or_clause($conditions));
 		return $Criteria;
 	}
 
-	protected function getSumaryAndTotal($works) {
+	protected function makeSumaryAndTotal() {
+		$works = $this->getWorks();
 		$professionals = array();
 		$sumary = array();
 		$totals = array();
@@ -347,9 +356,6 @@ class ChargeData {
 			if (empty($professionals[$matter_code])) {
 				$professionals[$matter_code] = array();
 			}
-
-			// WHY?
-			//$work['duracion_incobrables'] = $this->doubtfulDuration($matter_code, $user_id);
 
 			$works[$i] = $work;
 
@@ -455,12 +461,9 @@ class ChargeData {
 			$totals[$key]['glosa_flatfee'] = Utiles::Decimal2GlosaHora($data['flatfee']);
 		}
 
-		return array($sumary, $totals, $professionals);
-	}
-
-	protected function makeSumaryAndTotal() {
-		$works = $this->getWorks();
-		list($this->sumary, $this->totals, $this->matter_sumary) = $this->getSumaryAndTotal($works);
+		$this->sumary = $sumary;
+		$this->totals = $totals;
+		$this->matter_sumary = $professionals;
 		$this->sumary_and_totals = true;
 	}